Change dtype in detection module

+++ packages/yolov5/yolov5/models/yolo.py
@@ -37,7 +37,7 @@ class Detect(nn.Module):
 
                 y = x[i].sigmoid()
                 y[..., 0:2] = ((y[..., 0:2] * 2. - 0.5 + self.grid[i].to(x[i].device)) * self.stride[i]).type(y.dtype)  # xy
-                y[..., 2:4] = (y[..., 2:4] * 2) ** 2 * self.anchor_grid[i]  # wh
+                y[..., 2:4] = ((y[..., 2:4] * 2) ** 2 * self.anchor_grid[i]).type(y.dtype)  # wh
                 z.append(y.view(bs, -1, self.no))
 
         return x if self.training else (torch.cat(z, 1), x)
